android : refactor button_manager{.cpp, .h} to input_manager
 Now that it also handles sensors
