+++ src/android/app/src/main/jni/emu_window/emu_window.cpp
@@ -13,9 +13,9 @@
 #include "common/logging/log.h"
 #include "core/settings.h"
 #include "input_common/main.h"
-#include "jni/button_manager.h"
 #include "jni/emu_window/emu_window.h"
 #include "jni/id_cache.h"
+#include "jni/input_manager.h"
 #include "network/network.h"
 #include "video_core/renderer_base.h"
 #include "video_core/video_core.h"
