+++ src/android/app/src/main/jni/input_manager.cpp
@@ -14,7 +14,7 @@
 #include "common/param_package.h"
 #include "input_common/main.h"
 #include "input_common/sdl/sdl.h"
-#include "jni/button_manager.h"
+#include "jni/input_manager.h"
 #include "jni/ndk_motion.h"
 
 namespace InputManager {
