+++ src/android/app/src/main/jni/input_manager.h
